refactor: consolidate output blinding files
